% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/bambu-processReads_utilityConstructReadClasses.R
\name{assignGeneIds}
\alias{assignGeneIds}
\title{Main function which calls other gene id assigner functions
Returns a list of gene ids for each read class and an 
index of which genes are novel}
\usage{
assignGeneIds(
  grl,
  annotations,
  min.exonOverlap = 10,
  fusionMode = FALSE,
  maxChainIteration = 3
)
}
\arguments{
\item{grl}{a GrangesList object with read classes}

\item{annotations}{a GrangesList object with annotations}

\item{min.exonOverlap}{minimum length of overlap to be assigned to a gene}

\item{fusionMode}{if TRUE will assign multiple GENEIDs to fusion transcripts, separated by ":"}

\item{maxChainIteration}{the number of intermediate novel transcripts which will be used to assign gene ids, default: 3}
}
\description{
Main function which calls other gene id assigner functions
Returns a list of gene ids for each read class and an 
index of which genes are novel
}
