% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/bambu-processReads_utilityConstructReadClasses.R
\name{assignGeneIdsByReference}
\alias{assignGeneIdsByReference}
\title{Return gene ids for read classes which overlap
with known annotations}
\usage{
assignGeneIdsByReference(
  grl,
  annotations,
  min.exonOverlap = 10,
  fusionMode = FALSE,
  prefix = "Bambu"
)
}
\arguments{
\item{grl}{a GrangesList object with read classes}

\item{annotations}{a GrangesList object with annotations}
}
\description{
Return gene ids for read classes which overlap
with known annotations
}
