% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/bambu-processReads_utilityConstructReadClasses.R
\name{assignGeneIdsNonAssigned}
\alias{assignGeneIdsNonAssigned}
\title{Interal function which groups read classes that overlap}
\usage{
assignGeneIdsNonAssigned(
  geneTxMap,
  exonTxMap,
  geneExonMap,
  exonGeneMap_filter1,
  newExonId
)
}
\description{
Interal function which groups read classes that overlap
}
