% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/bambu-processReads_utilityConstructReadClasses.R
\name{assignGeneIdsNoReference}
\alias{assignGeneIdsNoReference}
\title{Create new gene ids for groups of overlapping read classes which 
don't overlap with known annotations.}
\usage{
assignGeneIdsNoReference(grl, prefix = "Bambu")
}
\arguments{
\item{grl}{a GrangesList object with read classes}
}
\description{
Create new gene ids for groups of overlapping read classes which 
don't overlap with known annotations.
}
