% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bambu-processReads_scoreReadClasses.R
\name{calculateGeneProportion}
\alias{calculateGeneProportion}
\title{% of a genes read counts assigned to each read class}
\usage{
calculateGeneProportion(counts, geneIds)
}
\description{
% of a genes read counts assigned to each read class
}
