% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bambu-processReads_scoreReadClasses.R
\name{checkFeatures}
\alias{checkFeatures}
\title{ensures that the data is trainable after filtering}
\usage{
checkFeatures(features, verbose = FALSE)
}
\description{
ensures that the data is trainable after filtering
}
