% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bambu-processReads_scoreReadClasses.R
\name{countPolyATerminals}
\alias{countPolyATerminals}
\title{returns number of A/T's each read class aligned 5' and 3' end}
\usage{
countPolyATerminals(grl, genomeSequence)
}
\description{
returns number of A/T's each read class aligned 5' and 3' end
}
