% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/bambu-processReads_utilityConstructReadClasses.R
\name{getRefGeneExonList}
\alias{getRefGeneExonList}
\title{Small function to meet bioconductor formatting requirements}
\usage{
getRefGeneExonList(exonTxMap, geneExonMap)
}
\description{
Small function to meet bioconductor formatting requirements
}
