% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bambu-processReads_scoreReadClasses.R
\name{getTranscriptScore}
\alias{getTranscriptScore}
\title{calculates a score based on how likely a read class is full length}
\usage{
getTranscriptScore(rowData, model = NULL, defaultModels)
}
\description{
calculates a score based on how likely a read class is full length
}
