% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bambu-processReads_scoreReadClasses.R
\name{isReadClassCompatible}
\alias{isReadClassCompatible}
\title{returns number of ref anno each read class is a subset of}
\usage{
isReadClassCompatible(query, subject)
}
\description{
returns number of ref anno each read class is a subset of
}
