% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bambu-extendAnnotations-utilityCombine.R
\name{makeUnsplicedTibble}
\alias{makeUnsplicedTibble}
\title{make unspliced tibble}
\usage{
makeUnsplicedTibble(
  combinedNewUnsplicedSe,
  newUnsplicedSeList,
  colDataNames,
  min.readCount,
  min.readFractionByGene,
  min.txScore.multiExon,
  min.txScore.singleExon,
  bpParameters
)
}
\description{
make unspliced tibble
}
