% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bambu-processReads_scoreReadClasses.R
\name{prepareTranscriptModelFeatures}
\alias{prepareTranscriptModelFeatures}
\title{calculate and format read class features for model training}
\usage{
prepareTranscriptModelFeatures(rowData)
}
\description{
calculate and format read class features for model training
}
