% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bambu-extendAnnotations-utilityExtend.R
\name{recommendNDR}
\alias{recommendNDR}
\title{calculates an expected NDR based on the annotations'}
\usage{
recommendNDR(
  combinedTranscripts,
  baselineFDR = 0.1,
  NDR = NULL,
  defaultModels = defaultModels,
  verbose = FALSE
)
}
\description{
calculates an expected NDR based on the annotations'
}
