% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bambu-processReads_scoreReadClasses.R
\name{scoreReadClasses}
\alias{scoreReadClasses}
\title{Assigns each read class geneScore and txScore}
\usage{
scoreReadClasses(
  se,
  genomeSequence,
  annotations,
  defaultModels,
  fit = TRUE,
  returnModel = FALSE,
  min.readCount = 2,
  min.exonOverlap = 10,
  fusionMode = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{se}{summerized experiment object with read classes/ranges}

\item{genomeSequence}{genomeSequence}

\item{annotations}{GRangesList of annotations}
}
\description{
Assigns each read class geneScore and txScore
}
