% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEtoGTF.R
\name{SEtoGTF}
\alias{SEtoGTF}
\title{transcript to gene expression}
\usage{
SEtoGTF(se)
}
\arguments{
\item{se}{a summarizedExperiment object from \code{\link{bambu}}}
}
\value{
gtf a GTF dataframe
}
\description{
Outputs a GTF dataframe for the nanorna-bam nextflow pipeline
}
