% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcriptToGeneExpression.R
\name{transcriptToGeneExpression}
\alias{transcriptToGeneExpression}
\title{transcript to gene expression}
\usage{
transcriptToGeneExpression(se)
}
\arguments{
\item{se}{a summarizedExperiment object from \code{\link{bambu}}}
}
\description{
Reduce transcript expression to gene expression
}
