% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWrite.R
\name{writeCountsOutput}
\alias{writeCountsOutput}
\title{helper function to write counts}
\usage{
writeCountsOutput(
  se,
  varname = "counts",
  feature = "transcript",
  outdir,
  prefix
)
}
\description{
helper function to write counts
}
