% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotationFunctions.R
\name{prepareAnnotations}
\alias{prepareAnnotations}
\title{prepare annotations from txdb object}
\usage{
prepareAnnotations(txdb)
}
\arguments{
\item{txdb}{a \code{\link{TxDb}} object}
}
\value{
A \code{\link{GrangesList}} object
}
\description{
Function to prepare tables and genomic ranges for transript reconstruction using a txdb object
}
\examples{
\dontrun{
 library(TxDb.Hsapiens.UCSC.hg38.knownGene)
 txdb <- TxDb.Hsapiens.UCSC.hg38.knownGene
 prepareAnnotations(txdb)
 }
}
