% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareAnnotations.R
\name{prepareAnnotations}
\alias{prepareAnnotations}
\title{prepare annotations from path to gtf file or txdb object}
\usage{
prepareAnnotations(x)
}
\arguments{
\item{x}{A path to gtf file or a \code{TxDb} object}
}
\value{
A \code{GRangesList} object
}
\description{
This function takes a transcript annotation (path to gtf file/txdb object) and extends the metadata for each transcript.
}
\details{
For each transcript, its exons are ranked in the \code{exon_rank} column according to the strandedness and the transcript coordinates of the transcript. The metadata also tells which gene each transcript (tx) belongs to and its transcript equivalence class. A transcript equivalence class of a tx is the collection of transcripts where their exon junctions contain, in a continuous way, the exon junctions of the tx. 
}
\examples{
gtf.file <- system.file("extdata",
    "Homo_sapiens.GRCh38.91_chr9_1_1000000.gtf",
    package = "bambu"
)
gtf_annotated <- prepareAnnotations(x = gtf.file)
gtf_annotated 

metadata <- mcols(gtf_annotated)
metadata
}
