% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotationFunctions.R
\name{prepareAnnotations}
\alias{prepareAnnotations}
\title{prepare annotations from txdb object or gtf file}
\usage{
prepareAnnotations(x)
}
\arguments{
\item{x}{A \code{\link{TxDb}} object or a gtf file}
}
\value{
A \code{\link{GRangesList}} object
}
\description{
Function to prepare tables and genomic ranges for transript reconstruction using a txdb object
}
\examples{
 gtf.file <- system.file("extdata", 
 "Homo_sapiens.GRCh38.91_chr9_1_1000000.gtf", 
 package = "bambu")
 prepareAnnotations(x = gtf.file)
}
