% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotationFunctions.R
\name{prepareAnnotationsFromGTF}
\alias{prepareAnnotationsFromGTF}
\title{Prepare annotation granges object from GTF file  into a GRangesList object}
\usage{
prepareAnnotationsFromGTF(file)
}
\arguments{
\item{file}{a GTF file}
}
\value{
A \code{\link{GRangesList}} object
}
\description{
Prepare annotation granges object from GTF file
}
\details{

}
\examples{
gtf.file <- system.file("extdata", "Homo_sapiens.GRCh38.91_chr9_1_1000000.gtf", package = "bambu")
gr <- prepareAnnotationsFromGTF(gtf.file)
}
