% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotationFunctions.R
\name{prepareAnnotationsFromGTF}
\alias{prepareAnnotationsFromGTF}
\title{Prepare annotation granges object from GTF file  into a \code{\link{GRangesList}}}
\usage{
prepareAnnotationsFromGTF(file)
}
\arguments{
\item{file}{a GTF file}
}
\value{

}
\description{
Prepare annotation granges object from GTF file
}
