% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWrite.R
\name{readFromGTF}
\alias{readFromGTF}
\title{convert a GTF file into a GRangesList}
\usage{
readFromGTF(file)
}
\arguments{
\item{file}{a \code{\link{.gtf}} file}
}
\value{
grlist a \code{\link{GRangesList}} object, with two columns
\itemize{
  \item TXNAME specifying prefix for new gene Ids (genePrefix.number), defaults to empty
  \item GENEID indicating whether filter to remove read classes which are a subset of known transcripts(), defaults to TRUE
  }
}
\description{
Outputs GRangesList object from reading a GTF file
}
