% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWrite.R
\name{writeBambuOutput}
\alias{writeBambuOutput}
\title{Write bambu results to GTF and transcript/gene-count files}
\usage{
writeBambuOutput(se, path)
}
\arguments{
\item{se}{a summarizedExperiment object from \code{\link{bambu}}}

\item{path}{the destination of the output files (gtf, transcript counts, and gene counts)}
}
\value{
The function will generate three files, a \code{\link{.gtf}} file for the annotations, 
two \code{\link{.txt}} files for transcript and gene counts respectively.
}
\description{
Outputs a GTF file, transcript-count file, and gene-count file from bambu
}
